jooby apt: generate source code #2968
- move classes to proper package: `apt`
