+++ modules/jooby-apt/src/main/java/io/jooby/internal/apt/MvcRoute.java
@@ -3,7 +3,7 @@
  * Apache License Version 2.0 https://jooby.io/LICENSE.txt
  * Copyright 2014 Edgar Espina
  */
-package io.jooby.internal.newapt;
+package io.jooby.internal.apt;
 
 import static java.util.Optional.ofNullable;
 
@@ -19,7 +19,6 @@ import com.squareup.javapoet.MethodSpec;
 import com.squareup.javapoet.ParameterSpec;
 import com.squareup.javapoet.TypeName;
 import io.jooby.apt.MvcContext;
-import io.jooby.internal.apt.Annotations;
 
 public class MvcRoute {
   private final MvcContext context;
