+++ modules/jooby-apt/src/main/java/io/jooby/internal/apt/MvcRouter.java
@@ -3,7 +3,7 @@
  * Apache License Version 2.0 https://jooby.io/LICENSE.txt
  * Copyright 2014 Edgar Espina
  */
-package io.jooby.internal.newapt;
+package io.jooby.internal.apt;
 
 import static java.util.Collections.emptyList;
