+++ modules/jooby-apt/src/main/java/io/jooby/internal/apt/TypeDefinition.java
@@ -3,7 +3,7 @@
  * Apache License Version 2.0 https://jooby.io/LICENSE.txt
  * Copyright 2014 Edgar Espina
  */
-package io.jooby.internal.newapt;
+package io.jooby.internal.apt;
 
 import java.util.ArrayList;
 import java.util.Collections;
