+++ modules/jooby-apt/src/test/java/tests/HandlerCompilerTest.java
@@ -35,10 +35,10 @@ import io.jooby.Formdata;
 import io.jooby.Reified;
 import io.jooby.Route;
 import io.jooby.StatusCode;
+import io.jooby.apt.MockContextHelper;
 import io.jooby.apt.NewProcessorRunner;
 import io.jooby.exception.MissingValueException;
 import io.jooby.exception.ProvisioningException;
-import io.jooby.internal.apt.MockContextHelper;
 import io.jooby.test.MockContext;
 import io.jooby.test.MockRouter;
 import source.CustomGenericType;
